% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_with_data.R
\name{cal_fit}
\alias{cal_fit}
\title{fitting cluster/perturbation model}
\usage{
cal_fit(data_list, fit_type)
}
\arguments{
\item{data_list}{the perturbation intensity, 0 means no perturbed, 1 means fully perturbed}

\item{n_species}{number of species}
}
\value{
fitted parameters
}
\description{
To fit the cluster/perturbation model and return the fitted parameters
}
