% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negbin_model.R
\name{cal_fit_negbin}
\alias{cal_fit_negbin}
\title{Fit data to negative binomial clustering model}
\usage{
cal_fit_negbin(data_list)
}
\arguments{
\item{data_list}{a list containing number of species, number of steps, number of clusters,
number of persons, the abundance data}
}
\value{
a list containing the sequence of states, centers and phi
}
\description{
Fit data to negative binomial clustering model
}
