% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_with_data.R
\name{cal_fit_perturbation}
\alias{cal_fit_perturbation}
\title{fitting perturbation model}
\usage{
cal_fit_perturbation(data_list)
}
\arguments{
\item{data_list}{data_list used in the model}
}
\value{
fitted parameters
}
\description{
To fit the perturbation model and return the fitted parameters
}
