% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_test.R
\name{estimate_power_clust}
\alias{estimate_power_clust}
\title{calculate the powers}
\usage{
estimate_power_clust(
  config,
  n_reps,
  simulator,
  ts_matrix,
  inital_state,
  centers
)
}
\arguments{
\item{config}{configurations}

\item{n_reps}{number of repetitions}

\item{simulator}{simulator we use}

\item{ts_matrix}{transition matrix}

\item{centers}{centroid of cluster}
}
\value{
power statistics
}
\description{
The function calculates the statistics of power test
}
