% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_test.R
\name{estimate_power_perturb}
\alias{estimate_power_perturb}
\title{calculate the powers}
\usage{
estimate_power_perturb(
  config,
  centroids,
  n_reps,
  simulator,
  sig_level,
  sim_opts
)
}
\arguments{
\item{config}{configurations}

\item{centroids}{centroids of the cluster}

\item{n_reps}{number of repetitions}

\item{simulator}{simulator we use}

\item{sig_level}{significance level}

\item{sim_opts}{true perturbed species}
}
\value{
power statistics
}
\description{
The function calculates the statistics of power test
}
