% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_test.R
\name{estimate_stat_mult}
\alias{estimate_stat_mult}
\title{calculate the powers}
\usage{
estimate_stat_mult(
  config,
  n_reps,
  ts_matrix,
  initial_state,
  centers,
  person_effect
)
}
\arguments{
\item{config}{configurations}

\item{n_reps}{number of repetitions}

\item{ts_matrix}{transition matrix}

\item{initial_state}{initial state}

\item{centers}{centroid of cluster}

\item{person_effect}{difference among persons in the simulation}
}
\value{
power statistics
}
\description{
The function calculates the statistics of power test
}
