% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negbin_model.R
\name{estimate_stat_negbin}
\alias{estimate_stat_negbin}
\title{Conduct power tests and estimate the power of each configuration}
\usage{
estimate_stat_negbin(
  config,
  n_reps,
  n_species,
  ts_matrix,
  initial_state,
  centers,
  phi
)
}
\arguments{
\item{config}{configurations with user defined number of timepoints and number of samples}

\item{n_reps}{number of replicates for each experimental configurations}

\item{n_species}{number of species}

\item{ts_matrix}{transition matrix}

\item{initial_state}{a vector containing the initial states of each person}

\item{centers}{the centroids of clusters}

\item{phi}{dispersion parameter of the negative bionomial model}
}
\description{
Conduct power tests and estimate the power of each configuration
}
