% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomial_model.R
\name{estimate_transitions}
\alias{estimate_transitions}
\title{Estimate a transition matrix from the given sequence of states}
\usage{
estimate_transitions(clusters)
}
\arguments{
\item{clusters}{the alpha sequence of states}
}
\description{
Estimate a transition matrix from the given sequence of states
}
