% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_test.R
\name{get_centroids}
\alias{get_centroids}
\title{Get the centroids of the fitted model}
\usage{
get_centroids(data_list)
}
\arguments{
\item{data_list}{list used in the fitting list(alpha=alpha, n_species=ncol(obs)-1, n_timepoints = length(alpha), y = obs))}
}
\value{
fitted centroids (beta)
}
\description{
The function get the centroids of the fitted model
}
