% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator_validation.R
\name{heat_map}
\alias{heat_map}
\title{Compare simulated with real data through heatmap}
\usage{
heat_map(sim_data, obs_data, relative_abundance, log_transform)
}
\arguments{
\item{sim_data}{simulated data}

\item{relative_abundance}{compare the relative_abundance or not}

\item{log_transform}{log transfrom or not}

\item{obs_datas}{observed data}
}
\description{
heatmap to compare the structure and abundance of simulated and real data.
}
