% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_test.R
\name{make_alphas}
\alias{make_alphas}
\title{Alpha sequence generator}
\usage{
make_alphas(n_timepoints, intervals, signal_size)
}
\arguments{
\item{n_timepoints}{number of time points}

\item{signal_size}{the maximum perturbation intensity (between 0 and 1)}

\item{interval}{the interval of perturbation (interval\link{1}-interval\link{2}): increase the permutation intensity.(interval\link{2}-interval\link{3}): achieve maximized perturbation intensity. (interval\link{3}-interval\link{4}): decrease the permutation intensity.}
}
\value{
alpha the perturbation intensity seuqence, 0 means no perturbed, 1 means fully perturbed
}
\description{
The function generate alpha (perturbation intensity sequence) which can be used in the simulator
}
