% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_test.R
\name{make_configurations_mult}
\alias{make_configurations_mult}
\title{configurations generator for multinomial model}
\usage{
make_configurations_mult(n_person, n_depth, n_timepoints)
}
\arguments{
\item{n_person}{the number of persons}

\item{n_depth}{sequence depth}

\item{n_timepoints}{number of time points}
}
\value{
configurations the configurations used in the power test
}
\description{
The function generate all configurations used in the power test
}
