% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_test.R
\name{make_configurations_negbin}
\alias{make_configurations_negbin}
\title{Make experimental configurations for power tests}
\usage{
make_configurations_negbin(n_timepoints, n_person)
}
\arguments{
\item{n_timepoints}{number of timepoints for each person}

\item{n_person}{number of persons in the experiment}
}
\description{
Make experimental configurations for power tests for negative bionomial model
}
