% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_without_data.R
\name{markov_sample}
\alias{markov_sample}
\title{simulate the markov chain from transition matrix}
\usage{
markov_sample(prob_matrix, n_steps, initial_state)
}
\description{
simulate the markov chain from transition matrix
}
