% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_test.R
\name{modify_centroids_fraction}
\alias{modify_centroids_fraction}
\title{Modify centroids}
\usage{
modify_centroids_fraction(centroids, fraction_perturbed = NULL)
}
\arguments{
\item{centroids}{the centorids of the cluster}
}
\value{
modified centroids
}
\description{
The function modifies the centroids to perturbed/unperturbed
}
