% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negbin_model.R
\name{normDESeq2}
\alias{normDESeq2}
\title{Data normalization}
\usage{
normDESeq2(physeq, whichOTUs = NULL, method = c("poscounts", "ratio"))
}
\description{
Data normalization
}
\references{
https://genomebiology.biomedcentral.com/articles/10.1186/s13059-020-02104-1#MOESM1
}
