% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_test.R
\name{power_matrix_transform_clust}
\alias{power_matrix_transform_clust}
\title{transform the matrix to the dataframe}
\usage{
power_matrix_transform_clust(matrix)
}
\arguments{
\item{matrix}{matrix of the power statistics}
}
\value{
transformed dataframe
}
\description{
The function transform the power statistics to the dataframe
}
