% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_test.R
\name{relative_abundence}
\alias{relative_abundence}
\title{calculate the relative abundance}
\usage{
relative_abundence(matrix)
}
\arguments{
\item{matrix}{the abundance data}
}
\value{
return the relative abundance
}
\description{
calculate the relative abundance
}
