% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomial_model.R
\name{sim_clust_multinomial}
\alias{sim_clust_multinomial}
\title{Simulate clustering data given a set of user defined parameters and model parameters from fitting existing data}
\usage{
sim_clust_multinomial(
  n_depth,
  n_species,
  ts_matrix,
  n_person,
  count,
  initial_state = NULL,
  centers,
  sigma,
  person_effect
)
}
\arguments{
\item{n_depth}{sequence depth}

\item{n_species}{number of species}

\item{ts_matrix}{transition matrix of the markov-chain}

\item{n_person}{number of persons}

\item{count}{a vector of count of samples per person}

\item{initial_state}{initial state of the markov-chain}

\item{centers}{the centroids of each cluster}

\item{sigma}{variance added inside the cluster}

\item{person_effect}{difference among persons in the simulation}
}
\value{
the simulated data
}
\description{
Simulate clustering data given a set of user defined parameters and model parameters from fitting existing data
}
