% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negbin_model.R
\name{sim_clust_negbin}
\alias{sim_clust_negbin}
\title{Simulate clustering data given a set of user defined parameters and model parameters from fitting existing data through negative bionomial model}
\usage{
sim_clust_negbin(
  n_species,
  ts_matrix,
  n_timepoints,
  initial_state,
  centers,
  phi
)
}
\arguments{
\item{n_species}{number of species}

\item{ts_matrix}{transition matrix of the markov-chain}

\item{initial_state}{initial state of the markov-chain}

\item{centers}{the centroids of each cluster}

\item{n_person}{number of persons}

\item{count}{a vector of count of samples per person}

\item{sigma}{variance added inside the cluster}

\item{person_effect}{difference among persons in the simulation}
}
\value{
the simulated data
}
\description{
Simulate clustering data given a set of user defined parameters and model parameters from fitting existing data through negative bionomial model
}
