% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_with_data.R
\name{sim_clust_obs}
\alias{sim_clust_obs}
\title{Simulate microbiome sample data through cluster model}
\usage{
sim_clust_obs(
  prob_matrix = NULL,
  initial_state = 1,
  n_clust,
  n_steps = NULL,
  n_depth = NULL,
  n_person,
  obs
)
}
\arguments{
\item{prob_matrix}{transition matrix of the markov-chain}

\item{initial_state}{initial state of the markov-chain}

\item{n_clust}{number of clusters}

\item{n_steps}{number of timepoints}

\item{n_depth}{sequence depth}

\item{n_person}{number of persons in the pilot data}

\item{obs}{pilot data}
}
\value{
the simulated data
}
\description{
Simulate microbiome abundance count data with user-specified number of samples, species and sequencing depth
}
\examples{
sim_clust_obs(initial_state = 1, n_clust =4, n_person = 1, obs=preg_data)
}
