% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_with_data.R
\name{sim_perturb_obs}
\alias{sim_perturb_obs}
\title{Simulate microbiome sample data through perturbation model}
\usage{
sim_perturb_obs(alpha, obs, n_depth = NULL)
}
\arguments{
\item{alpha}{the perturbation intensity, 0 means no perturbed, 1 means fully perturbed}

\item{obs}{observed microbiome data}

\item{n_depth}{sequence depth}
}
\value{
the simulated data
}
\description{
Simulate microbiome abundance count data with user-specified number of samples, species and sequencing depth
}
\examples{
sim_perturb_obs(alpha = c(0,0,0.8,0.8,0.8,0,0,0,0,0), obs = data, n_depth = 2000)
}
