% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_without_data.R
\name{sim_samples}
\alias{sim_samples}
\title{Simulate microbiome sample data}
\usage{
sim_samples(n_samples, n_species, n_depth, Sigma = NULL)
}
\arguments{
\item{n_samples}{number of samples}

\item{n_species}{number of species}

\item{n_depth}{sequence depth}

\item{Sigma}{correlation matrix for muti-nomal distribution}
}
\value{
the simulated data
}
\description{
Simulate microbiome abundance count data with user-specified number of samples, species and sequencing depth
}
\examples{
simulate_samples(100,100,1000)
}
