% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_without_data.R
\name{sim_ts}
\alias{sim_ts}
\title{Simulate microbiome time series data}
\usage{
sim_ts(endtime, n_species, n_depth, Sigma = NULL, phi)
}
\arguments{
\item{endtime}{the number of time points of the sample}

\item{n_species}{number of species}

\item{n_depth}{sequence depth}

\item{Sigma}{correlation matrix for muti-nomal distribution}

\item{phi}{the slope of the increasing}
}
\value{
the simulated data
}
\description{
simulate microbiome longitudinal count data for one sample
}
\examples{
simulate_ts(20,100,1000)
}
