% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_without_data.R
\name{sim_ts_cluster}
\alias{sim_ts_cluster}
\title{Simulate microbiome clustered data}
\usage{
sim_ts_cluster(
  ts_matrix,
  initial_state,
  n_sample,
  n_steps,
  n_species,
  n_depth,
  centers = NULL
)
}
\arguments{
\item{ts_matrix}{transition matrix of markov chain}

\item{initial_state}{the initial state of the markov chain}

\item{n_sample}{number of samples}

\item{n_steps}{number of time points}

\item{n_species}{number of species}

\item{n_depth}{sequence depth}
}
\value{
the simulated data
}
\description{
To simulate microbiome longitudinal count data with clustering and markov chain based on user input transition matrix
}
\examples{
sim_ts_cluster(ts_matrix=matrix(c(0.70, 0.2,0.1,0.3,0.4,0.3,0,0.4,0.6), nrow=3, byrow=TRUE), n_steps=25, initial_state = 3, n_species=100, n_depth=1000)
}
