% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_without_data.R
\name{sim_ts_perturb_2}
\alias{sim_ts_perturb_2}
\title{Simulate microbiome perturbation data using clusted model}
\usage{
sim_ts_perturb_2(
  alpha,
  n_species = 30,
  Sigma = NULL,
  n_depth,
  effect,
  beta = NULL
)
}
\arguments{
\item{alpha}{the perturbation intensity, 0 means no perturbed, 1 means fully perturbed}

\item{n_species}{number of species}

\item{Sigma}{variance of different speicies in the same cluster}

\item{n_depth}{sequence depth}

\item{effect}{the effect of perturbation}

\item{beta}{the centroids of perturbed/unperturbed cluster}
}
\value{
the simulated data
}
\description{
To simulate microbiome longitudinal count data for perturbation experiments with user defined alpha, effect, and centroids of perturbed/unperturbed data
}
\examples{
stan_sim(alpha=alpha, n_depth = 1000, n_species = 30, Sigma=NULL, effect=0)
}
