% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator_validation.R
\name{specieswise_plot}
\alias{specieswise_plot}
\title{Compare simulated with real data}
\usage{
specieswise_plot(sim_data, obs_data, relative = FALSE)
}
\arguments{
\item{sim_data}{simulated data}

\item{relative}{compare the relative_abundance or not}

\item{obs_datas}{observed data}
}
\description{
species-wise scatter plot to compare the abundance of simulated and real data.
}
