% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator_validation.R
\name{violin_plot}
\alias{violin_plot}
\title{Compare simulated with real data}
\usage{
violin_plot(sim_data, obs_data, relative = FALSE)
}
\arguments{
\item{sim_data}{simulated data}

\item{relative}{compare the relative_abundance or not}

\item{obs_datas}{observed data}
}
\description{
Violin plot to compare the mean and variance of simulated and real data.
}
