% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomial_model.R
\name{cal_fit_clust_multinomial}
\alias{cal_fit_clust_multinomial}
\title{Fit the lnm model and out the centroids of cluster}
\usage{
cal_fit_clust_multinomial(data_list)
}
\arguments{
\item{data_list}{a list containing number of species, number of clusters, pilot data
number of persons, and a vector containing number of samples for each person}
}
\value{
a list with the markov-chain sequence, the centroids of clusters, variance, persons'effect, the transition matrix
}
\description{
Fit the lnm model and out the centroids of cluster
}
