% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_test.R
\name{make_configurations_clust}
\alias{make_configurations_clust}
\title{configurations generator for clust}
\usage{
make_configurations_clust(ns, n_depth, n_timepoints)
}
\arguments{
\item{ns}{the number of samples}

\item{n_depth}{sequence depth}

\item{n_timepoints}{number of time points}

\item{alpha}{the maximum perturbation intensity (between 0 and 1)}

\item{ts_matrix}{transition matrix}
}
\value{
configurations the configurations used in the power test
}
\description{
The function generate all configurations used in the power test
}
