% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomial_model.R,
%   R/simulation_without_data.R
\name{markov_sample}
\alias{markov_sample}
\title{Generate a sequence of state}
\usage{
markov_sample(prob_matrix, n_steps, initial_state)

markov_sample(prob_matrix, n_steps, initial_state)
}
\arguments{
\item{prob_matrix}{transition matrix among clusters}

\item{n_steps}{number of steps in teh markov chain}

\item{initial_state}{initial states for each person}
}
\value{
the generated markov-chain
}
\description{
Generate a sequence of states from the given transition matrix
}
