% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator_validation.R
\name{relative_abundence}
\alias{relative_abundence}
\title{Calculate relative abundance}
\usage{
relative_abundence(matrix)
}
\arguments{
\item{matrix}{microbiome abundance data}
}
\value{
return the relative abundance
}
\description{
Calculate the relative abundance of the microbiome data
}
